feat: reusable checkbox component, rule content
