+++ src/components/ContainsRuleContent.vue
@@ -1,7 +1,9 @@
 <script setup>
 import {ref} from "vue";
+import CompactCheckbox from "./CompactCheckbox.vue";
 
 const contains = ref('');
+const caseSensitive = ref(false);
 
 </script>
 
@@ -9,6 +11,7 @@ const contains = ref('');
 <div>
   This will check each message to see if it contains the following input:
   <input type="text" v-model="contains" class="input__contains" placeholder="Contains..." />
+  <CompactCheckbox label="Case Sensitive" v-model="caseSensitive"/>
 </div>
 </template>
 
@@ -19,8 +22,8 @@ const contains = ref('');
   border-radius: 0.25rem;
   border: 1px solid var(--bg4);
   background-color: var(--bg2);
-  font-size: 0.95em;
-  margin-top: 0.5rem;
+  margin-top: 0.75rem;
+  margin-bottom: 0.5rem;
   color: var(--text);
 }
 </style>
