+++ src/components/ExactMatchRuleContent.vue
@@ -1,7 +1,9 @@
 <script setup>
 import {ref} from "vue";
+import CompactCheckbox from "./CompactCheckbox.vue";
 
 const contains = ref('');
+const caseSensitive = ref(false);
 
 </script>
 
@@ -9,7 +11,7 @@ const contains = ref('');
 <div>
   This will check each message to see if it exactly matches the following input:
   <input type="text" v-model="contains" class="input__contains" placeholder="Contains..." />
-  <v-checkbox v-model="caseSensitive" label="Case Sensitive" class="checkbox__case-sensitive"></v-checkbox>
+  <CompactCheckbox label="Case Sensitive" v-model="caseSensitive"/>
 </div>
 </template>
 
@@ -20,8 +22,9 @@ const contains = ref('');
   border-radius: 0.25rem;
   border: 1px solid var(--bg4);
   background-color: var(--bg2);
-  font-size: 0.95em;
-  margin-top: 0.5rem;
+  font-size: 1em;
+  margin-top: 0.75rem;
+  margin-bottom: 0.5rem;
   color: var(--text);
 }
 </style>
